라우팅 구조 개편 (#549)
* chore: layout 태그 변경

* refactor: layout 컴포넌트 -> ui 제외 로직 분리

* feat: appRoutes 정적 파일 생성

* feat: RootElement 컴포넌트 생성 (과거 App)

* refactor: 데스크탑 및 모바일 routes 제거

* refactor: App 파일 제거

* refactor: 서비스 진입점 수정

* refactor: appRoutes 홈 element 분기처리 수정
