+++ frontend/src/index.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
-import '@/styles/index.css';
+import { createBrowserRouter, RouterProvider } from 'react-router';
+import appRoutes from './routes';
 import * as Sentry from '@sentry/react';
 import GTM from '@/libs/trackers/GTM';
+import '@/styles/index.css';
 
 if (process.env.NODE_ENV === 'production') {
   Sentry.init({
@@ -15,9 +16,10 @@ if (process.env.NODE_ENV === 'production') {
 }
 
 const root = ReactDOM.createRoot(document.getElementById('root')!);
+const router = createBrowserRouter(appRoutes);
 
 root.render(
   <React.StrictMode>
-    <App />
+    <RouterProvider router={router} />;
   </React.StrictMode>
 );
