+++ frontend/src/shared/layout/Layout.styled.ts
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-export const Container = styled.div`
+export const Container = styled.main`
   width: 100vw;
   min-height: calc(100vh - 5rem - 150px);
   background-color: ${({ theme }) => ` ${theme.colors.gray05}`};
