+++ frontend/src/shared/layout/index.tsx
@@ -1,9 +1,6 @@
 import * as S from './Layout.styled';
 import Header from './Header';
 import { Outlet } from 'react-router';
-import ErrorBoundary from '@/providers/ErrorBoundary';
-import ErrorPage from '@/pages/common/ErrorPage';
-import ToastProvider from '@/providers/ToastProvider';
 import Footer from './Footer';
 import { useTheme } from '@emotion/react';
 
@@ -13,13 +10,9 @@ const Layout = ({ isDark, toggleTheme }: { isDark: boolean; toggleTheme: () => v
   return (
     <>
       <Header isDark={isDark} toggleTheme={toggleTheme} />
-      <ToastProvider>
-        <ErrorBoundary fallback={<ErrorPage />}>
-          <S.Container>
-            <Outlet />
-          </S.Container>
-        </ErrorBoundary>
-      </ToastProvider>
+      <S.Container>
+        <Outlet />
+      </S.Container>
       {!isMobile && <Footer />}
     </>
   );
